--- a/crates/cairo-lang-executable/src/compile.rs
+++ b/crates/cairo-lang-executable/src/compile.rs
@@ -4,7 +4,7 @@ use anyhow::{Context, Result};
 use cairo_lang_compiler::db::RootDatabase;
 use cairo_lang_compiler::diagnostics::DiagnosticsReporter;
 use cairo_lang_compiler::project::setup_project;
-use cairo_lang_compiler::{DbWarmupContext, get_sierra_program_for_functions};
+use cairo_lang_compiler::{ensure_diagnostics, get_sierra_program_for_functions};
 use cairo_lang_debug::debug::DebugWithDb;
 use cairo_lang_executable_plugin::{
     EXECUTABLE_PREFIX, EXECUTABLE_RAW_ATTR, executable_plugin_suite,
@@ -128,8 +128,7 @@ pub fn compile_executable_in_prepared_db<'db>(
     mut diagnostics_reporter: DiagnosticsReporter<'_>,
     config: ExecutableConfig,
 ) -> Result<CompileExecutableResult<'db>> {
-    let context = DbWarmupContext::new();
-    context.ensure_diagnostics(db, &mut diagnostics_reporter)?;
+    ensure_diagnostics(db, &mut diagnostics_reporter)?;
 
     let executables = find_executable_functions(db, main_crate_ids, executable_path);
 
@@ -152,7 +151,7 @@ pub fn compile_executable_in_prepared_db<'db>(
         }
     };
 
-    compile_executable_function_in_prepared_db(db, executable, config, context)
+    compile_executable_function_in_prepared_db(db, executable, config)
 }
 
 /// Search crates identified by `main_crate_ids` for executable functions.
@@ -209,10 +208,9 @@ pub fn compile_executable_function_in_prepared_db<'db>(
     db: &'db dyn Database,
     executable: ConcreteFunctionWithBodyId<'db>,
     config: ExecutableConfig,
-    context: DbWarmupContext,
 ) -> Result<CompileExecutableResult<'db>> {
     let SierraProgramWithDebug { program: sierra_program, debug_info } =
-        get_sierra_program_for_functions(db, vec![executable], context)
+        get_sierra_program_for_functions(db, vec![executable])
             .with_context(|| "Compilation failed without any diagnostics.")?;
     if !config.allow_syscalls {
         // Finding if any syscall libfuncs are used in the program.