With the new salsa, I don't see any benefit from having the limit anymore. 


OZ benchmarks:

Before:

Benchmark 1: SCARB_INCREMENTAL=false scarb build -w --test
  Time (mean ± σ):     106.724 s ±  4.403 s    [User: 313.918 s, System: 10.279 s]
  Range (min … max):   103.255 s … 111.678 s    3 runs

After:

Benchmark 1: SCARB_INCREMENTAL=false scarb build -w --test
  Time (mean ± σ):     92.257 s ±  2.130 s    [User: 420.124 s, System: 22.883 s]
  Range (min … max):   90.839 s … 94.707 s    3 runs